build: upgrade to latest platform constructs pipeline
