+++ test/util.ts
@@ -1,0 +1,26 @@
+import { Aspects, Stack } from 'aws-cdk-lib';
+import { Annotations, Match } from 'aws-cdk-lib/assertions';
+import { AwsSolutionsChecks } from 'cdk-nag';
+import { synthesisMessageToString } from '@orcabus/platform-cdk-constructs/utils';
+
+/**
+ * Run the CDK nag checks.
+ */
+export function cdkNagStack(stack: Stack, applySuppressions: (stack: Stack) => void) {
+  Aspects.of(stack).add(new AwsSolutionsChecks());
+  applySuppressions(stack);
+
+  test(`cdk-nag AwsSolutions Pack errors`, () => {
+    const errors = Annotations.fromStack(stack)
+      .findError('*', Match.stringLikeRegexp('AwsSolutions-.*'))
+      .map(synthesisMessageToString);
+    expect(errors).toHaveLength(0);
+  });
+
+  test(`cdk-nag AwsSolutions Pack warnings`, () => {
+    const warnings = Annotations.fromStack(stack)
+      .findWarning('*', Match.stringLikeRegexp('AwsSolutions-.*'))
+      .map(synthesisMessageToString);
+    expect(warnings).toHaveLength(0);
+  });
+}
