+++ tsconfig.json
@@ -25,6 +25,10 @@
     "typeRoots": [
       "./node_modules/@types"
     ],
+    "types": [
+      "node",
+      "jest"
+    ],
     "skipLibCheck": true
   },
   "exclude": [
